Create helpers and conftest for python tests
