+++ .devcontainer/python_development/devcontainer.json
@@ -10,7 +10,8 @@
     "vscode": {
       "extensions": [
         "ms-python.python",
-        "littlefoxteam.vscode-python-test-adapter"
+        "littlefoxteam.vscode-python-test-adapter",
+        "ms-python.black-formatter"
       ]
     }
   }
