+++ .devcontainer/python_tests/devcontainer.json
@@ -1,15 +1,18 @@
 {
-    "name": "Python Tests",
-    "dockerComposeFile": ["../../testing/tests/docker-compose.yml"],
-    "service": "ubuntu",
-    "shutdownAction": "none",
-    "workspaceFolder": "/app",
-    "customizations": {
-        "vscode": {
-          "extensions": [
-            "ms-python.python",
-            "littlefoxteam.vscode-python-test-adapter"
-          ]
-        }
-      }
+  "name": "Python Tests",
+  "dockerComposeFile": [
+    "../../testing/tests/docker-compose.yml"
+  ],
+  "service": "ubuntu",
+  "shutdownAction": "none",
+  "workspaceFolder": "/app",
+  "customizations": {
+    "vscode": {
+      "extensions": [
+        "ms-python.python",
+        "littlefoxteam.vscode-python-test-adapter",
+        "ms-python.black-formatter"
+      ]
+    }
+  }
 }
