+++ testing/tests/README.md
@@ -16,6 +16,50 @@ You can copy the versions in the `testing/tests/.vscode` folder, as a starting p
 
 Using Docker helps to avoid polluting your host environment with multiple versions of Python.
 
+### Running tests in the Development Container Option
+When you select the Python Tests option to run your container in, there are already
+config files for running tests in VSCode so you won't have to set this part up. 
+
+If you want to run tests within the 
+Python Development environment option, you will have to make a `.vscode/launch.json` in the root 
+of your environment. This folder isn't checked into the repo so it has to be manually
+created. 
+The easy way to create this file is to click on the play button (triangle) with the little bug on it in your 
+VSCode activity bar. There will be a link there to "create a launch.json file". Click on that link and select 
+"Python Debugger"->"Python File". This will create a file and open it. Replace its contents with the below 
+code to run the `api_tests` in `testing/tests/api_tests`.
+After that, the Run and Debug interface will change and have a green arrow in it for running and testing code. 
+
+```
+{
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "name": "Python Debugger: Run Tests",
+            "type": "debugpy",
+            "request": "launch",
+            "module": "pytest",
+            "args": [
+                "${workspaceFolder}/testing/tests/api_tests" // Path to your tests
+            ],
+            "console": "integratedTerminal",
+            "justMyCode": false, // Set this to false to allow debugging into external libraries
+            "cwd": "${workspaceFolder}/testing/tests/" // Set the working directory
+        }
+    ]
+}
+```
+If you want to get the test explorer (beaker icon) to be able to find your tests, you can add
+this to your `.vscode/settings.json`, so it knows to look in the `/testing/tests` folder. 
+```
+"python.testing.pytestArgs": [
+    "testing/tests"
+],
+"python.testing.unittestEnabled": false,
+"python.testing.nosetestsEnabled": false,
+"python.testing.pytestEnabled": true
+```
+
 ## VSCode Extensions
 The necessary VSCode extensions have been installed, in the Python Tests container, for
 running and debugging tests within VSCode. The first time you open the project in a
@@ -24,12 +68,17 @@ container, it may take a little time for VSCode to install the necessary extensi
 ## Environment Variables Setup
 - There is an example `.env_example` file for setting environment variables for the tests.
 - To use it, copy this file and rename it to `.env`.
-- The testing environment will then pick up those variables and set them as environment variables before running tests.
+- The testing environment will then pick up those variables and set them as environment 
+variables before running tests.
 
 ## Python Virtual Environment Setup
 In order for VSCode to use the python modules for the tests, you will want to install a
-python virtual environment for it to use. 
-You can do this by opening a new terminal in VSCode, within the `testing/tests` directory, and running:
+python virtual environment for it to use. You can make a python virtual environment
+folder that is available for both of the development containers by making it in the 
+`testing/tests` folder. Then you can have only one copy of the environment for both 
+container options. 
+You can do this by opening a new terminal in VSCode, within the `testing/tests` 
+directory, and running:
 
 
 `python3 -m venv venv`
